+++ lib/core/hive.dart/hive_impl.dart
@@ -1,12 +1,11 @@
 import 'dart:io';
 
 import 'package:ejyption_time_2/core/contacts/contacts_impl_flutter_contacts.dart';
-import 'package:ejyption_time_2/core/contacts/contacts_provider_interface.dart';
 import 'package:ejyption_time_2/models/global/global_model.dart';
 import 'package:ejyption_time_2/models/meeting/meeting.dart';
 import 'package:hive/hive.dart';
 
-import 'package:ejyption_time_2/core/hive.dart/hive_cipher_impl.dart';
+//import 'package:ejyption_time_2/core/hive.dart/hive_cipher_impl.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:path_provider/path_provider.dart';
 
@@ -33,48 +32,30 @@ class HiveImpl {
   Future<void> openAllBoxes() async {
     Directory directory = await getApplicationDocumentsDirectory();
     BoxCollection collection = await BoxCollection.open(
-      'EgyptionTime', // Name of your database
+      'EgyptionTime',
       {
         'Meetings',
         'PointAssestments',
         'ProbabilityAssestments'
       }, // Names of your boxes
-      path: directory.path
-          .toString(), // Path where to store your boxes (Only used in Flutter / Dart IO)
+      path: directory.path.toString(),
       // key:
-      //     HiveCipherImpl(), // Key to encrypt your boxes (Only used in Flutter / Dart IO)
+      //     HiveCipherImpl(), // toDo: enctiprion
     );
     meetingBox = (await collection.openBox<String>('Meetings'));
   }
-
-  // Future<Map<bool, List<Meeting>?>> updateMeetingList() async {
-  //   if (permissionDenied || meetingBox == null) {
-  //     return {false: null};
-  //   } else {
-  //     return meetingBox!.getAllValues().then((value) =>
-  //         {true: value.entries.map((e) => Meeting.fromJson(e.value)).toList()});
-  //   }
-
-  //   //  {GlobalModel.instance.meetings.updateMeetingList(
-  //   //     MeetingBox.getAllValues().then((value) => .entries
-  //   //         .map((e) => Meeting.fromJson(e.value))
-  //   //         .toList(),)
-
-  //   //     false);}
-  // }
 
   Future<void> updateMeetingList() async {
     if (meetingBox == null) {
       return;
     }
-    GlobalModel.instance.meetings.updateMeetingList(
-        (await meetingBox!.getAllValues())
-            .entries
-            .map((e) => Meeting.fromJson(e.value))
-            .toList(),
-        false);
-    GlobalModel.instance.meetings.listIsUpdating = false;
-    GlobalModel.instance.meetings.changeModel();
+    List<Meeting> retrievedMettings = await retrieveAllMeetings();
+    GlobalModel.instance.meetings.updateMeetingList(retrievedMettings, false);
+  }
+
+  retrieveAllMeetings() async {
+    return meetingBox!.getAllValues().then((rawMap) =>
+        rawMap.entries.map((e) => Meeting.fromJson(e.value)).toList());
   }
 
   saveMeetingList() async {
@@ -86,6 +67,4 @@ class HiveImpl {
   Future<void> remove(id) async {
     meetingBox?.delete(id);
   }
-  // Open your boxes. Optional: Give it a type.
-
 }
