+++ lib/core/shared/const.dart
@@ -1,0 +1,1 @@
+const bool isMyTest = false;
