+++ lib/ui/features/detailed_meeting/meeting_cover_over_detailed.dart
